fix parsing of backslash at end of line
If you have an option with a line ending with a backslash that's on the
last line of the file (or preceeding the last line when it's a comment
or empty space), then the option isn't added to the corresponding
section. I discovered this because the [k3s installer](https://get.k3s.io/)
generates a unit file that looks like this:
```
[Service]
Type=notify
ExecStart=/usr/local/bin/k3s \
        server \
```

but attempting to start it generates the error `ERROR:systemctl: k3s.service: Service lacks both ExecStart and ExecStop= setting. Refusing.`

Signed-off-by: Mason Malone <651224+MasonM@users.noreply.github.com>